python/requirements.txt
anyio==3.5.0
asgiref==3.5.0
bidict==0.22.0
certifi==2021.10.8
click==8.1.0
fastapi==0.75.0
h11==0.13.0
idna==3.3
msgpack==1.0.3
protobuf==3.19.4
pydantic==1.9.0
python-engineio==4.3.1
python-socketio==5.5.2
ray==1.11.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
